Moved TextdrawCreator commands to separate class
